refactor(mean.independent.noninferiority): Extract the common `_power_xxx` function
